1)) Fixed ensureDatePickersAreCollapesed so that it removes the correct expand class based on the index of the element. 2)) Made the calendarClickHandler clear the endDate field if the startDate is greater than the endDate 3)) Added the colour blue to dates on the datepicker that are clickable 4)) Made the date picker for endDate forbid choosing an endDate in the past and coloured all the forbidden tds red. 5)) Did some minor moving around of the methods in taskform.js placing dependent function next to each other
A note about forEach: Changed the code in forEach to vary depending on the index. Originally the project was simple and all of the elements I was using forEach on needed the same handling functions but with added features such as error handling, the requirements changed. However I wanted the forEach method to remain so I switched to relying on the new scope created by forEach to get the index. I understand that I am attaching a function with the code for all the indexes on all the elements of the nodeList. I am simply doing this to experiment and to improve my understanding of scope.